Update project configuration and dependencies
The .vscode/settings.json file was removed since the automatic update build configuration for Java was no longer necessary. The pom.xml has been adjusted to add the parent configuration for Spring Boot and to use the parent's version instead of specifying it separately. This simplifies the project setup and makes it easier to manage the Spring Boot version. Also, the Spring Cloud version property has been added to track the version of Spring Cloud dependencies. The plugin management section is replaced with the plugins section, with the spring boot maven plugin version now derived from parent too, making the whole configuration more concise and in line with the Spring Boot's recommended best practices.