Add product-service and set up initial configurations
This commit introduces a new microservice, product-service, into the project. It includes the setup of initial dependency configurations in the pom.xml of the product-service and parent pom.xml. To ensure proper version control and isolation of development environment, files and directories such as .idea and target are added to .gitignore. For testing, ProductServiceApplicationTests and TestProductServiceApplication are created with an embedded MySQL test container. A minimalist Product model is also added to start building the main functionalities of the service. Docker compose file is added for container orchestration. Lastly, ProductServiceApplication class is introduced as the main entry point for the product-service application.