"Add DTO mapping to ProductService"
Added ProductMapper to ProductService to transform Product entities into DTOs. The changes were made to create a separation between the database entities and the objects that are returned from the API endpoints. ProductService's findAll method now returns a ProductPageDTO containing a list of ProductDTOs, replacing the previous return type of a Page of Products.