Refactor test structure and configuration
The ProductServiceApplicationTests class was deleted and replaced with ProductServiceApplicationTest and TestProductServiceApplication was renamed to TestContainersConfig for clarity and structure. A new dependency 'spring-boot-devtools' was added in the test scope of pom.xml, this will provide automatic restarts which is useful during development. New TestClasses 'TestApplication' and 'BaseIntegrationTest' were added to increase test coverage and capability.