Add findById method and handle NotFoundException
This commit introduces a new method (findById) in the ProductService and ProductController classes to allow fetching a product by its ID. It also includes a new ProductNotFoundException class and a @RestControllerAdvice (ApplicationControllerAdvice) to handle this new exception. These changes are made to improve the API's error handling and expand its functionality, in response to requirements for being able to retrieve individual products by their unique IDs. Tests have been added to ensure the correct functioning of these new features.