Add Status enum and converter to Product
Added a Status enum (Active, Inactive) to standardize the product status values. Converted the 'status' attribute in the Product.class from String to the new Status enum type. Introduced a database converter to change the enum back and forth to a String for storage in the database. These changes bring more type safety, thus reducing the potential for errors.