Refactor tests configuration & Add InventoryController
Refactored the test configuration class name from 'TestInventoryServiceApplication' to 'TestContainersConfig', to better reflect its purpose. The test main method was removed as it is not needed in this context.

Added 'InventoryControllerTest' to prepare for application testing with real MongoDB instance provided by TestContainers. This automatically cleans database after each test.

Also, new 'InventoryController' was created to handle requests for inventory data, to support the application's new functionality.

'InventoryController' relies on 'InventoryService', which is to be implemented.

Added 'TestApplication' with main method running tests with TestContainersConfig class.