Configure Discovery and Gateway services
Updated 'application.properties' files for both Discovery and Gateway Services. The purpose of these changes is to set up the basic configurations needed for these services to operate correctly in a Eureka server setup. In the discovery service, we have defined Eureka specific properties and set it to not register itself with Eureka. In Gateway