Integrate InventoryClient into ProductController
Added InventoryClient into ProductController to fetch product inventory information. Now, when finding a product by id, the service will also return the quantity of the product available in the inventory. This InventoryClient is connected to the "inventory-service" via Feign. Also, the Inventory object was created to map the response received from the remote call.